chore: source code organization
Organization
- Tools are now moved to independent folders named similarly for structure
- Tools are now bundled in the same folder with their controllers
