+++ lib/exports/barrel.dart
@@ -25,16 +25,16 @@ export 'dart:io';
 export 'dart:typed_data';
 export 'package:academia/controllers/courses_page_controller.dart';
 export 'package:academia/controllers/dashboard_controller.dart';
-export 'package:academia/pages/attendance_page.dart';
+export 'package:academia/tools/attendance_page/attendance_page.dart';
 export 'package:flutter_carousel_widget/flutter_carousel_widget.dart';
 export 'package:connectivity_plus/connectivity_plus.dart';
-export 'package:academia/pages/exams_timetable_page.dart';
-export 'package:academia/pages/fees_page.dart';
-export 'package:academia/pages/gpacalculator_page.dart';
+export 'package:academia/tools/exam_timetable/exams_timetable_page.dart';
+export 'package:academia/tools/fees/fees_page.dart';
 export 'package:academia/pages/pdf_viewer.dart';
 export 'package:academia/pages/tasks_pages/taskmanager_page.dart';
 export 'package:academia/pages/webview_page.dart';
-export 'package:academia/controllers/gpacalculator_controller.dart';
+export 'package:academia/tools/gpa_calculator/gpacalculator_controller.dart';
+export 'package:academia/tools/gpa_calculator/gpacalculator_page.dart';
 export 'package:academia/pages/courses_page.dart';
 export 'package:academia/pages/dashboard.dart';
 export 'package:academia/pages/settings_page.dart';
@@ -46,9 +46,11 @@ export 'package:academia/widgets/info_card.dart';
 export 'package:cached_network_image/cached_network_image.dart';
 export 'package:academia/widgets/semester_timeline_tile.dart';
 export 'package:academia/constants/tools.dart';
-export 'package:academia/pages/birthday_page.dart';
+export 'package:academia/tools/birthday/birthday_page.dart';
 export 'package:academia/widgets/tool_card.dart';
 export 'package:academia/pages/tasks_pages/edittask_page.dart';
 export 'package:academia/pages/tasks_pages/newtask_page.dart';
 export 'package:academia/pages/tasks_pages/taskinfo_page.dart';
 export 'package:magnet/src/magnet_utils.dart';
+export 'package:academia/tools/time_line/time_line_page.dart';
+export 'package:academia/tools/stories/notifications_story_page.dart';
